pyproject.toml
[tool.poetry]
name = "temporal-practice-ground"
version = "0.1.0"
description = ""
authors = ["deight93 <deight8311@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.6"
fastapi = "^0.110.3"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.29"
alembic = "^1.13.1"
pre-commit = "^3.7.0"
pydantic-settings = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
